requirements.txt
discord.py==2.3.2
transformers==4.36.2
torch==2.1.0
sentencepiece==0.1.99
python-dotenv==1.0.0
protobuf==4.25.1
flask==3.0.0